shadow-native: Disable syslog usage
Spamming the build host's syslog with useradd information during image creation
isn't great. Add a patch to disable that.

It may be possible to convince upstream to make it a configure option but for
now the patch is trivial and reduces host impact to the logs.

(From OE-Core rev: 9e3b2a8ff2f33163be6f1cc0608f6973c0b57561)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>